requirements.txt
customtkinter==5.2.2
keyboard==0.13.5
Pillow==9.5.0
pynput==1.7.7
pyperclip==1.8.2
pystray==0.19.5
tencentcloud_sdk_python_common==3.0.1214
tencentcloud_sdk_python_tmt==3.0.1214
python==3.9.16